requirements.txt
PyMuPDF>=1.24.10
transformers>=4.42.0
torch>=2.1.0
sentencepiece>=0.2.0
accelerate>=0.33.0
numpy>=1.26.0
scikit-learn>=1.4.0
bert-extractive-summarizer>=0.10.1
sentence-transformers>=3.0.1
pytesseract>=0.3.10
Pillow>=10.3.0
streamlit>=1.36.0
tqdm>=4.66.0
nltk>=3.9
